+++ src/kinegrant/experimental/__init__.py
@@ -1,0 +1,1 @@
+"""Non-normative profiles used to build reproducible KGP experiments."""